+++ packages/terra-framework-docs/CHANGELOG.md
@@ -14,6 +14,9 @@
 * Added
   * Added the 'aria-label' attribute to announce the information for the 'Flowsheet-Data Grid'.
 
+* Changed
+  * Updated `terra-folder-tree` test example.
+
 ## 1.62.0 - (January 26, 2024)
 
 * Added