+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.64.0 - (February 7, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 1.63.0 - (February 1, 2024)
 
 * Changed